Create Video Games
The global market for video games is expected to reach more than $300 billion in value by 2027. The major game publishers are very motivated to produce the next big hit because there is so much money on the line. But are games improving, or has the peak of their popularity already passed?

We'll examine the top 400 video games released from 1977 and 2020 in this study. To assess whether or whether video games have gotten better as the gaming industry has expanded, we'll compare a dataset on game sales with user and critic reviews.

Two tables make up our database. For this project, we've restricted each table to 400 rows, however Kaggle has the whole dataset with more than 13,000 games.